+++ vkworkspace/utils/__init__.py
@@ -1,9 +1,35 @@
 from .keyboard import InlineKeyboardBuilder, InlineKeyboardButton
-from .text import html, md, split_text
+from .text import (
+    Bold,
+    Code,
+    Italic,
+    Link,
+    Mention,
+    Pre,
+    Quote,
+    Raw,
+    Strikethrough,
+    Text,
+    Underline,
+    html,
+    md,
+    split_text,
+)
 
 __all__ = [
+    "Bold",
+    "Code",
     "InlineKeyboardBuilder",
     "InlineKeyboardButton",
+    "Italic",
+    "Link",
+    "Mention",
+    "Pre",
+    "Quote",
+    "Raw",
+    "Strikethrough",
+    "Text",
+    "Underline",
     "html",
     "md",
     "split_text",